backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Firebase
firebase-admin==6.4.0

# Validation and utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# CORS
starlette==0.35.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Production server
gunicorn==21.2.0
